clean up ArtistsGrid component
